version  1.7

+++ Api/DTOs/ProductoDto.cs
@@ -9,7 +9,7 @@ namespace Api.DTOs
     {
         public  string Id { get; set; }
         public  string Nombre { get; set; }
-        public int Precio { get; set; }
+        public float Precio { get; set; }
         public string Departamento { get; set; }
         public string Tipo { get; set; }
         public string Descripcion { get; set; }
